Move Graph component to graph visualization directory
