+++ .eslintrc.json
@@ -17,7 +17,8 @@
     "react/no-deprecated": "warn",
     "react/display-name": "warn",
     "react/no-unescaped-entities": "off",
-    "import/no-webpack-loader-syntax": "off"
+    "import/no-webpack-loader-syntax": "off",
+    "no-restricted-imports": ["error", { "patterns": ["*graphVisualization*"] }]
   },
   "globals": {
     "it": true,
@@ -89,7 +90,7 @@
           "error",
           {
             "paths": ["redux", "react-redux"],
-            "patterns": ["browser*"]
+            "patterns": ["browser*", "shared*", "services*"]
           }
         ]
       }
