+++ src/graphVisualization/components/WheelZoomInfoOverlay.tsx
@@ -1,5 +1,26 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 import React from 'react'
 
+import { isMac } from '../utils/utils'
+import { InfoIcon } from './Icons'
 import {
   StyledZoomInfo,
   StyledZoomInfoIconContainer,
@@ -8,8 +29,6 @@ import {
   StyledZoomInfoText,
   StyledZoomInfoTextContainer
 } from './styled'
-import { InfoIcon } from 'browser-components/icons/Icons'
-import { isMac } from 'shared/utils/platformUtils'
 
 const getModKeyString = () => (isMac ? 'Cmd' : 'Ctrl or Shift')
