+++ src/graphVisualization/index.ts
@@ -1,16 +1,33 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+export { Graph, GraphProps } from './components/Graph'
+
 export { GraphModel } from './models/Graph'
+export { GetNodeNeighboursFn } from './models/GraphEventHandler'
 export { GraphStyleModel, Selector } from './models/GraphStyle'
-export { NodeModel } from './models/Node'
-export {
-  RelationshipModel,
-  RelationshipCaptionLayout
-} from './models/Relationship'
 
+export { NodeItem, RelationshipItem, VizItem, VizItemProperty } from './types'
+
+export {
+  GraphStats,
+  GraphStatsLabels,
+  GraphStatsRelationshipTypes
+} from './utils/mapper'
 export { measureText } from './utils/textMeasurement'
-export {
-  MeasureSizeFn,
-  Visualization,
-  ZoomLimitsReached
-} from './visualization/Visualization'
-
-export * from './types'
