+++ src/graphVisualization/models/GraphStyle.ts
@@ -17,10 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import {
-  selectorArrayToString,
-  selectorStringToArray
-} from 'services/grassUtils'
+import { selectorArrayToString, selectorStringToArray } from '../utils/utils'
 
 export class Selector {
   tag = ''
