+++ src/shared/services/grassUtils.test.ts
@@ -17,12 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import {
-  objToCss,
-  parseGrass,
-  selectorArrayToString,
-  selectorStringToArray
-} from 'services/grassUtils'
+import { objToCss, parseGrass } from 'services/grassUtils'
 
 describe('parseGrass', () => {
   it('should create an object from a valid CSS string', () => {
@@ -138,44 +133,5 @@ relationship {
   it('does not break on undefined', () => {
     const css = objToCss()
     expect(css).toEqual(false)
-  })
-})
-
-describe('selectors', () => {
-  test('selectorStringToArray unescapes selectors with . in them', () => {
-    const tests = [
-      { test: 'foo', expect: ['foo'] },
-      { test: 'foo.bar', expect: ['foo', 'bar'] },
-      { test: 'foo.bar.baz', expect: ['foo', 'bar', 'baz'] },
-      { test: 'foo\\.bar', expect: ['foo.bar'] },
-      { test: 'foo\\.bar\\.baz', expect: ['foo.bar.baz'] },
-      { test: 'foo\\.bar.baz\\.baz', expect: ['foo.bar', 'baz.baz'] },
-      {
-        test: 'node.foo\\.bar\\.baz.bax',
-        expect: ['node', 'foo.bar.baz', 'bax']
-      }
-    ]
-
-    tests.forEach(t => {
-      expect(selectorStringToArray(t.test)).toEqual(t.expect)
-    })
-  })
-  test('selectorArrayToString escapes selectors with . in them', () => {
-    const tests = [
-      { expect: 'foo', test: ['foo'] },
-      { expect: 'foo.bar', test: ['foo', 'bar'] },
-      { expect: 'foo.bar.baz', test: ['foo', 'bar', 'baz'] },
-      { expect: 'foo\\.bar', test: ['foo.bar'] },
-      { expect: 'foo\\.bar\\.baz', test: ['foo.bar.baz'] },
-      { expect: 'foo\\.bar.baz\\.baz', test: ['foo.bar', 'baz.baz'] },
-      {
-        expect: 'node.foo\\.bar\\.baz.bax',
-        test: ['node', 'foo.bar.baz', 'bax']
-      }
-    ]
-
-    tests.forEach(t => {
-      expect(selectorArrayToString(t.test)).toEqual(t.expect)
-    })
   })
 })
